dom/tests/mochitest/dom-level1-core/test_nodereplacechildnewchildexists.html

<!DOCTYPE HTML PUBLIC "-//W3C//DTD HTML 4.01//EN""http://www.w3.org/TR/html4/strict.dtd">
<html>
<head>
<META http-equiv="Content-Type" content="text/html; charset=UTF-8">
<meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
<title>http://www.w3.org/2001/DOM-Test-Suite/level1/core/nodereplacechildnewchildexists</title>
<link href="http://www.w3.org/StyleSheets/activity-home.css" rel="stylesheet" type="text/css">
<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
<script type="text/javascript" src="/MochiKit/packed.js"></script><script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script><script type="text/javascript" src="DOMTestCase.js"></script><script type="text/javascript" src="exclusions.js"></script><script type="text/javascript">
// expose test function names
function exposeTestFunctionNames()
{
return ['nodereplacechildnewchildexists'];
}

var docsLoaded = -1000000;
var builder = null;

//
//   This function is called by the testing framework before
//      running the test suite.
//
//   If there are no configuration exceptions, asynchronous
//        document loading is started.  Otherwise, the status
//        is set to complete and the exception is immediately
//        raised when entering the body of the test.
//
function setUpPage() {
   setUpPageStatus = 'running';
   try {
     //
     //   creates test document builder, may throw exception
     //
     builder = createConfiguredBuilder();

      docsLoaded = 0;
      
      var docRef = null;
      if (typeof(this.doc) != 'undefined') {
        docRef = this.doc;
      }
      docsLoaded += preload(docRef, "doc", "staff");
        
       if (docsLoaded == 1) {
          setUpPage = 'complete';
       }
    } catch(ex) {
    	catchInitializationError(builder, ex);
        setUpPage = 'complete';
    }
}

//
//   This method is called on the completion of 
//      each asychronous load started in setUpTests.
//
//   When every synchronous loaded document has completed,
//      the page status is changed which allows the
//      body of the test to be executed.
function loadComplete() {
  if (++docsLoaded == 1) {
    setUpPageStatus = 'complete';
    runJSUnitTests();
    markTodos();
    SimpleTest.finish();
  }
}

var docName = 'nodereplacechildnewchildexists';



window.doc = window;  
SimpleTest.waitForExplicitFinish();
addLoadEvent(setUpPage);
function testFails (test) {
  if (!test.result) {
    test.todo = true;
    return true;
  }
  return false;
}
function markTodos() {
  if (todoTests[docName]) {
    // mark the failures as todos
    var failures = filter(testFails, SimpleTest._tests);
    // shouldn't be 0 failures
    todo(SimpleTest._tests != 0 && failures == 0, "test marked todo should fail somewhere");
  }
}

function runJSUnitTests() {
  builder = createConfiguredBuilder();
  try {
   forEach(exposeTestFunctionNames(), 
     function (testName) { 
       window[testName](); 
     }
   );
  } catch (ex) {
    ok(false, "Test threw exception: " + ex);
  }
}


/**
* 
    Retrieve the second employee and replace its TWELFTH 
    child(address) with its SECOND child(employeeId).   After the
    replacement the second child should now be the one that used   
    to be at the third position and the TWELFTH child should be the
    one that used to be at the SECOND position.

* @author NIST
* @author Mary Brady
* @see http://www.w3.org/TR/1998/REC-DOM-Level-1-19981001/level-one-core#ID-785887307
*/
function nodereplacechildnewchildexists() {
   var success;
    if(checkInitialization(builder, "nodereplacechildnewchildexists") != null) return;
    var doc;
      var elementList;
      var employeeNode;
      var childList;
      var oldChild = null;

      var newChild = null;

      var childName;
      var childNode;
      var length;
      var actual = new Array();

      var expected = new Array();

      expectedWithoutWhitespace = new Array();
      expectedWithoutWhitespace[0] = "name";
      expectedWithoutWhitespace[1] = "position";
      expectedWithoutWhitespace[2] = "salary";
      expectedWithoutWhitespace[3] = "gender";
      expectedWithoutWhitespace[4] = "employeeId";

      expectedWithWhitespace = new Array();
      expectedWithWhitespace[0] = "#text";
      expectedWithWhitespace[1] = "#text";
      expectedWithWhitespace[2] = "name";
      expectedWithWhitespace[3] = "#text";
      expectedWithWhitespace[4] = "position";
      expectedWithWhitespace[5] = "#text";
      expectedWithWhitespace[6] = "salary";
      expectedWithWhitespace[7] = "#text";
      expectedWithWhitespace[8] = "gender";
      expectedWithWhitespace[9] = "#text";
      expectedWithWhitespace[10] = "employeeId";
      expectedWithWhitespace[11] = "#text";

      var replacedChild;
      
      var docRef = null;
      if (typeof(this.doc) != 'undefined') {
        docRef = this.doc;
      }
      doc = load(docRef, "doc", "staff");
      elementList = doc.getElementsByTagName("employee");
      employeeNode = elementList.item(1);
      childList = employeeNode.childNodes;

      length = childList.length;

      
	if(
	(13 == length)
	) {
	newChild = childList.item(1);
      oldChild = childList.item(11);
      expected =  expectedWithWhitespace;

	}
	
		else {
			newChild = childList.item(0);
      oldChild = childList.item(5);
      expected =  expectedWithoutWhitespace;

		}
	replacedChild = employeeNode.replaceChild(newChild,oldChild);
      assertSame("return_value_same",oldChild,replacedChild);
for(var indexN65758 = 0;indexN65758 < childList.length; indexN65758++) {
      childNode = childList.item(indexN65758);
      childName = childNode.nodeName;

      actual[actual.length] = childName;

	}
   assertEqualsList("childNames",expected,actual);
       
}

</script>
</head>
<body>
<h2>Test http://www.w3.org/2001/DOM-Test-Suite/level1/core/nodereplacechildnewchildexists</h2>
<p></p>
<p>
Copyright (c) 2001-2004 World Wide Web Consortium, 
(Massachusetts Institute of Technology, European Research Consortium 
for Informatics and Mathematics, Keio University). All 
Rights Reserved. This work is distributed under the <a href="http://www.w3.org/Consortium/Legal/2002/copyright-software-20021231">W3C(r) Software License</a> in the 
hope that it will be useful, but WITHOUT ANY WARRANTY; without even 
the implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
</p>
</body>
</html>
